Update GestioneCatalogoServiceImpl.java
tipo parametro price da int a float

+++ Sistema/src/main/java/application/GestioneCatalogoService/GestioneCatalogoServiceImpl.java
@@ -302,7 +302,7 @@ public class GestioneCatalogoServiceImpl implements GestioneCatalogoService{
 	 * 
 	 * */
 	
-	public Collection<ProxyProdotto> aggiornamentoPrezzoProdotto(Prodotto product, int price, int page, int perPage) throws CategoriaProdottoException, SottocategoriaProdottoException, SQLException, ProdottoNonInCatalogoException, PrezzoProdottoException{
+	public Collection<ProxyProdotto> aggiornamentoPrezzoProdotto(Prodotto product, float price, int page, int perPage) throws CategoriaProdottoException, SottocategoriaProdottoException, SQLException, ProdottoNonInCatalogoException, PrezzoProdottoException{
 		ProdottoDAODataSource productDao = new ProdottoDAODataSource();
 		ProxyProdotto retrieved = productDao.doRetrieveProxyByKey(product.getCodiceProdotto());
